+++ package.json
@@ -18,6 +18,7 @@
     "@types/jest": "^26.0.14",
     "concurrently": "^5.3.0",
     "enzyme-adapter-react-16": "^1.15.6",
+    "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
     "jest": "21.x",
     "jest-resolve": "^26.4.0",
     "jsdom": "^16.6.0",
