+++ packages/module/src/ConsoleInternal/vendor.scss
@@ -8,6 +8,12 @@
 
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/normalize";
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/scaffolding";
-
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/type";
 @import "~bootstrap-sass/assets/stylesheets/bootstrap/code";
+
+// .co-quick-start-panel-content,
+// .ocs-page-layout__content {
+//     @import "~bootstrap-sass/assets/stylesheets/bootstrap/scaffolding";
+//     @import "~bootstrap-sass/assets/stylesheets/bootstrap/type";
+//     @import "~bootstrap-sass/assets/stylesheets/bootstrap/code";
+// }
