+++ packages/module/src/ConsoleShared/src/components/popper/index.ts
@@ -1,1 +1,2 @@
-export { default as Popper } from './Popper';
+export { default as SimplePopper } from './SimplePopper';
+export { default as Portal } from './Portal';
