+++ packages/module/src/ConsoleShared/src/components/status/StatusIconAndText.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import classNames from 'classnames';
 import { CamelCaseWrap } from '@console/internal/components/utils';
 import { DASH } from '../../constants';
 import { StatusComponentProps } from './types';
+import classNames from 'classnames';
 
 type StatusIconAndTextProps = StatusComponentProps & {
   icon?: React.ReactElement;
