+++ packages/module/src/ConsoleShared/src/components/status/statuses.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import { HourglassHalfIcon, InProgressIcon } from '@patternfly/react-icons';
 import {
   BlueInfoCircleIcon,
   GreenCheckCircleIcon,
   RedExclamationCircleIcon,
   YellowExclamationTriangleIcon,
 } from './icons';
-import GenericStatus from './GenericStatus';
+import { HourglassHalfIcon, InProgressIcon } from '@patternfly/react-icons';
 import { StatusComponentProps } from './types';
+import GenericStatus from './GenericStatus';
 
 export const ErrorStatus: React.FC<StatusComponentProps> = (props) => (
   <GenericStatus {...props} Icon={RedExclamationCircleIcon} title={props.title || 'Error'} />
