+++ packages/module/src/catalog/QuickStartTile.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { CatalogTile } from '@patternfly/react-catalog-view-extension';
-import RocketIcon from '@patternfly/react-icons/dist/esm/icons/rocket-icon';
 import { FallbackImg } from '@console/shared';
 import { QuickStart, QuickStartStatus } from '../utils/quick-start-types';
-import QuickStartTileHeader from './QuickStartTileHeader';
+import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
+import { camelize } from '../utils/quick-start-utils';
 import QuickStartTileDescription from './QuickStartTileDescription';
 import QuickStartTileFooter from './QuickStartTileFooter';
 import QuickStartTileFooterExternal from './QuickStartTileFooterExternal';
-import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
-import { camelize } from '../utils/quick-start-utils';
+import QuickStartTileHeader from './QuickStartTileHeader';
+import RocketIcon from '@patternfly/react-icons/dist/esm/icons/rocket-icon';
 
 import './QuickStartTile.scss';
