+++ packages/module/src/catalog/QuickStartTileFooterExternal.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Button, Flex, FlexItem } from '@patternfly/react-core';
-import ExternalLinkAltIcon from '@patternfly/react-icons/dist/esm/icons/external-link-alt-icon';
 import { QuickStartExternal } from '../utils/quick-start-types';
+import ExternalLinkAltIcon from '@patternfly/react-icons/dist/esm/icons/external-link-alt-icon';
 
 type QuickStartTileFooterProps = {
   link: QuickStartExternal;
