+++ packages/module/src/controller/QuickStartConclusion.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
 import { Button } from '@patternfly/react-core';
+import { QuickStart, QuickStartTask, QuickStartTaskStatus } from '../utils/quick-start-types';
+import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
 import ArrowRightIcon from '@patternfly/react-icons/dist/esm/icons/arrow-right-icon';
 import QuickStartMarkdownView from '../QuickStartMarkdownView';
-import { QuickStart, QuickStartTask, QuickStartTaskStatus } from '../utils/quick-start-types';
 import TaskHeader from './QuickStartTaskHeader';
 
 type QuickStartConclusionProps = {
