+++ packages/module/src/controller/QuickStartContent.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { QuickStart, QuickStartTaskStatus } from '../utils/quick-start-types';
+import QuickStartConclusion from './QuickStartConclusion';
 import QuickStartIntroduction from './QuickStartIntroduction';
 import QuickStartTasks from './QuickStartTasks';
-import QuickStartConclusion from './QuickStartConclusion';
 
 import './QuickStartContent.scss';
