+++ packages/module/src/controller/QuickStartIntroduction.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
-import QuickStartMarkdownView from '../QuickStartMarkdownView';
 import { QuickStartTask, QuickStartTaskStatus } from '../utils/quick-start-types';
+import QuickStartMarkdownView from '../QuickStartMarkdownView';
 import TaskHeader from './QuickStartTaskHeader';
 
 type QuickStartIntroductionProps = {
