+++ packages/module/src/controller/QuickStartTaskHeader.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
-import cx from 'classnames';
 import { QuickStartContext, QuickStartContextValues } from '../utils/quick-start-context';
-import { Title, WizardNavItem } from '@patternfly/react-core';
-import CheckCircleIcon from '@patternfly/react-icons/dist/esm/icons/check-circle-icon';
-import ExclamationCircleIcon from '@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon';
 import { QuickStartTaskStatus } from '../utils/quick-start-types';
+import { Title, WizardNavItem } from '@patternfly/react-core';
 import { markdownConvert } from '../ConsoleInternal/components/markdown-view';
 import { removeParagraphWrap } from '../QuickStartMarkdownView';
+import CheckCircleIcon from '@patternfly/react-icons/dist/esm/icons/check-circle-icon';
+import ExclamationCircleIcon from '@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon';
+import cx from 'classnames';
 
 import './QuickStartTaskHeader.scss';
