+++ packages/module/src/data/quick-start-test-data.ts
@@ -1,7 +1,7 @@
+import { QuickStart } from '@patternfly/quickstarts';
 import { explorePipelinesQuickStart } from './mocks/json/explore-pipeline-quickstart';
 import { exploreServerlessQuickStart } from './mocks/json/explore-serverless-quickstart';
 import { monitorSampleAppQuickStart } from './mocks/json/monitor-sampleapp-quickstart';
-import { QuickStart } from '@patternfly/quickstarts';
 
 export const allQuickStarts: QuickStart[] = [
   explorePipelinesQuickStart,
